Hoist mobile Voice mode logic into client-core useVoiceMode hook (#1213)
Move the framework-agnostic voice browser modules (clips.ts, playbackPositions.ts)
from apps/mobile/src/voice into packages/client-core/src/voice so both the mobile
app and the Cockpit can share them, and extract every piece of non-presentational
logic from the mobile VoiceMode page into a new shared useVoiceMode hook.

The mobile VoiceMode page is now a thin view: it reads the session id and the
roster's voice-mode seed from the router and renders byte-identical JSX over the
hook. No mobile behavior changes. The logic (the poll, the state-machine
derivation, the playback handlers, and the clip management) is moved verbatim.

Phase 4 of the Cockpit improvement plan. The Cockpit Voice tab will render over
this same hook.
